added mimic project structure, increased  expectations durations for some tests
