+++ FitpaySDKTests/BaseTestProvider.swift
@@ -1,0 +1,14 @@
+//
+//  BaseTestProvider.swift
+//  FitpaySDK
+//
+//  Created by Illya Kyznetsov on 5/29/18.
+//  Copyright © 2018 Fitpay. All rights reserved.
+//
+
+import XCTest
+@testable import FitpaySDK
+
+class BaseTestProvider: XCTestCase {
+    let mockModels = MockModels()
+}
