+++ FitpaySDKTests/FitpaySDKTests.swift
@@ -28,7 +28,7 @@ class FitpaySDKTests: XCTestCase {
     
     func testPerformanceExample() {
         // This is an example of a performance test case.
-        self.measureBlock {
+        self.measure {
             // Put the code you want to measure the time of here.
         }
     }
